requirements_test.txt
pytest-homeassistant-custom-component
pytest-asyncio
freezegun
